requirements.txt
streamlit==1.31.1
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
plotly>=5.18.0
scikit-learn==1.4.0
tensorflow>=2.15.0
scipy>=1.11.0
redis==5.0.1
websockets==12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
aiohttp==3.9.3
reportlab==4.1.0 